requirements.txt
openai>=1.0.0
anthropic>=0.7.0
feedparser>=6.0.0
requests>=2.28.0
pydub>=0.25.1
pyyaml>=6.0
python-dotenv>=1.0.0
openai-whisper>=20231117
whisperx>=3.1.1
pyannote-audio>=3.1.0
# For AMD ROCm support, install PyTorch with:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/rocm5.6
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0